+++ group-coordinator/src/test/java/org/apache/kafka/coordinator/group/GroupMetadataManagerTest.java
@@ -19370,2434 +19370,6 @@ public class GroupMetadataManagerTest {
     }
 
     @Test
-    public void testStaticMemberJoinThenRevokeAndReceiveTasks() {
-        int enoughMaxSize = 100;
-        testStaticMemberJoinThenRevokeAndReceiveTasksWith2Members(enoughMaxSize);
-    }
-
-    @Test
-    public void testStaticMemberJoinThenRevokeAndReceiveTasksInMaxSizeBoundary() {
-        int boundarySize = 2;
-        testStaticMemberJoinThenRevokeAndReceiveTasksWith2Members(boundarySize);
-    }
-
-    @Test
-    public void testStaticMemberRejoinWithUpdatedProcessIdBumpsStreamsGroupEpoch() {
-        String groupId = "fooup";
-        String oldMemberId = Uuid.randomUuid().toString();
-        String rejoinMemberId = Uuid.randomUuid().toString();
-        String instanceId = Uuid.randomUuid().toString();
-
-        String oldProcessId = "old-process-id";
-        String newProcessId = "new-process-id";
-        String subtopology1 = "subtopology1";
-        String fooTopicName = "foo";
-        Uuid fooTopicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopology1).setSourceTopics(List.of(fooTopicName))
-        ));
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(fooTopicId, fooTopicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                fooTopicName, computeTopicHash(fooTopicName, metadataImage)
-        ));
-
-        MockTaskAssignor assignor = new MockTaskAssignor("sticky");
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(assignor))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, 10)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(oldMemberId, instanceId)
-                                .setMemberEpoch(-2)
-                                .setPreviousMemberEpoch(10)
-                                .setProcessId(oldProcessId)
-                                .setAssignedTasks(mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, 10,
-                                        TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3)))
-                                .build())
-                        .withTargetAssignment(oldMemberId, TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE,
-                                TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3)))
-                        .withTargetAssignmentEpoch(10)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withMetadataHash(groupMetadataHash)
-                        .withValidatedTopologyEpoch(0)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs())
-                )
-                .build();
-
-        assignor.prepareGroupAssignment(Map.of(
-                rejoinMemberId, TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE,
-                        TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3))
-        ));
-
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> result = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(rejoinMemberId)
-                        .setMemberEpoch(0)
-                        .setProcessId(newProcessId)
-                        .setActiveTasks(List.of())
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of())
-        );
-
-        assertEquals(rejoinMemberId, result.response().data().memberId());
-        assertEquals(11, result.response().data().memberEpoch());
-
-        CoordinatorRecord metadataRecord = result.records().stream()
-                .filter(record -> record.key() instanceof StreamsGroupMetadataKey)
-                .findFirst()
-                .orElse(null);
-
-        assertNotNull(metadataRecord, "Expected a StreamsGroupMetadata record when static member config changes.");
-        StreamsGroupMetadataValue metadataValue = (StreamsGroupMetadataValue) metadataRecord.value().message();
-        assertEquals(11, metadataValue.epoch());
-
-        assertTrue(result.records().contains(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMetadataRecord(
-                        groupId,
-                        11,
-                        groupMetadataHash,
-                        0,
-                        getDefaultAssignmentConfigs()
-                )
-        ));
-    }
-
-    @Test
-    public void testStaticMemberLeaveWithMinusOneFencesMemberAndBumpsStreamsGroupEpoch() {
-        String groupId = "fooup";
-        String memberId = Uuid.randomUuid().toString();
-        String instanceId = Uuid.randomUuid().toString();
-
-        String subtopology1 = "subtopology1";
-        String fooTopicName = "foo";
-        Uuid fooTopicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopology1).setSourceTopics(List.of(fooTopicName))
-        ));
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(fooTopicId, fooTopicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                fooTopicName, computeTopicHash(fooTopicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, 10)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(memberId, instanceId)
-                                .setMemberEpoch(10)
-                                .setPreviousMemberEpoch(9)
-                                .setAssignedTasks(mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, 10,
-                                        TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3)))
-                                .build())
-                        .withTargetAssignment(memberId, TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE,
-                                TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3)))
-                        .withTargetAssignmentEpoch(10)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs())
-                )
-                .build();
-
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> result = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(memberId)
-                        .setMemberEpoch(LEAVE_GROUP_MEMBER_EPOCH)
-        );
-
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(memberId)
-                        .setMemberEpoch(LEAVE_GROUP_MEMBER_EPOCH)
-                        .setStatus(List.of()),
-                result.response().data()
-        );
-
-        assertRecordsEquals(
-                List.of(
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentTombstoneRecord(groupId, memberId),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentTombstoneRecord(groupId, memberId),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupMemberTombstoneRecord(groupId, memberId),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupMetadataRecord(
-                                groupId, 11, groupMetadataHash, 0, getDefaultAssignmentConfigs()),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentMetadataRecord(
-                                groupId, 11, 0L)
-                ),
-                result.records()
-        );
-    }
-
-    @Test
-    public void testStaticMemberLeaveWithLeaveGroupStaticMemberEpoch() {
-        // GIVEN
-        int leaveEpoch = LEAVE_GROUP_STATIC_MEMBER_EPOCH;
-        int memberEpoch = 10;
-        int groupEpoch = 10;
-
-        String groupId = "fooup";
-        String memberId = Uuid.randomUuid().toString();
-        String instanceId = Uuid.randomUuid().toString();
-
-        String subtopology1 = "subtopology1";
-        String fooTopicName = "foo";
-        Uuid fooTopicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopology1).setSourceTopics(List.of(fooTopicName))
-        ));
-
-        // GIVEN Tasks
-        Map.Entry<String, Set<Integer>> givenTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3);
-        TasksTupleWithEpochs givenAssignedTask = mkTasksTupleWithCommonEpoch(
-                TaskRole.ACTIVE, groupEpoch, givenTaskEntries
-        );
-        TasksTupleWithEpochs givenTasksPendingRevocation = mkTasksTupleWithCommonEpoch(
-                TaskRole.ACTIVE, groupEpoch, TaskAssignmentTestUtil.mkTasks(subtopology1, 2, 3)
-        );
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(fooTopicId, fooTopicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                fooTopicName, computeTopicHash(fooTopicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(memberId, instanceId)
-                                .setMemberEpoch(memberEpoch)
-                                .setPreviousMemberEpoch(memberEpoch - 1)
-                                .setAssignedTasks(givenAssignedTask)
-                                .setTasksPendingRevocation(givenTasksPendingRevocation)
-                                .build())
-                        .withTargetAssignment(memberId, TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, givenTaskEntries))
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs())
-                )
-                .build();
-
-        // WHEN
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> result = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(memberId)
-                        .setMemberEpoch(leaveEpoch)
-        );
-
-        // THEN
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(memberId)
-                        .setMemberEpoch(leaveEpoch)
-                        .setStatus(List.of()),
-                result.response().data()
-        );
-
-        // No group epoch bump.
-        // Member epoch should be -2.
-        // task still remain. 
-        // pendingRevocationTasks should be EMPTY.
-        StreamsGroupMember expectedMemberInResponse = streamsGroupMemberBuilderWithDefaults(memberId, instanceId)
-                .setMemberEpoch(leaveEpoch)
-                .setPreviousMemberEpoch(memberEpoch - 1)
-                .setTasksPendingRevocation(TasksTupleWithEpochs.EMPTY)
-                .setAssignedTasks(givenAssignedTask)
-                .build();
-        assertRecordsEquals(
-                List.of(StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, expectedMemberInResponse)),
-                result.records()
-        );
-        assertEquals(groupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-    }
-
-    @Test
-    public void testStaticMemberLeaveWithLeaveGroupStaticMemberEpochThenShouldBeIdempotence() {
-        // GIVEN
-        int leaveEpoch = LEAVE_GROUP_STATIC_MEMBER_EPOCH;
-        int memberEpoch = 10;
-        int groupEpoch = 10;
-
-        String groupId = "fooup";
-        String memberId = Uuid.randomUuid().toString();
-        String instanceId = Uuid.randomUuid().toString();
-
-        String subtopology1 = "subtopology1";
-        String fooTopicName = "foo";
-        Uuid fooTopicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopology1).setSourceTopics(List.of(fooTopicName))
-        ));
-
-        // GIVEN Tasks
-        Map.Entry<String, Set<Integer>> givenTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3);
-        TasksTupleWithEpochs givenAssignedTask = mkTasksTupleWithCommonEpoch(
-                TaskRole.ACTIVE, groupEpoch, givenTaskEntries
-        );
-        TasksTupleWithEpochs givenTasksPendingRevocation = mkTasksTupleWithCommonEpoch(
-                TaskRole.ACTIVE, groupEpoch, TaskAssignmentTestUtil.mkTasks(subtopology1, 2, 3)
-        );
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(fooTopicId, fooTopicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                fooTopicName, computeTopicHash(fooTopicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(memberId, instanceId)
-                                .setMemberEpoch(memberEpoch)
-                                .setPreviousMemberEpoch(memberEpoch - 1)
-                                .setAssignedTasks(givenAssignedTask)
-                                .setTasksPendingRevocation(givenTasksPendingRevocation)
-                                .build())
-                        .withTargetAssignment(memberId, TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, givenTaskEntries))
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs())
-                )
-                .build();
-
-        // GIVEN request
-        StreamsGroupHeartbeatRequestData request = new StreamsGroupHeartbeatRequestData()
-                .setGroupId(groupId)
-                .setInstanceId(instanceId)
-                .setMemberId(memberId)
-                .setMemberEpoch(leaveEpoch);
-
-        // Expect
-        StreamsGroupHeartbeatResponseData expectedResponseData = new StreamsGroupHeartbeatResponseData()
-                .setMemberId(memberId)
-                .setMemberEpoch(leaveEpoch)
-                .setStatus(List.of());
-
-        // No group epoch bump.
-        // Member epoch should be -2.
-        // task still remain. 
-        // pendingRevocationTasks should be EMPTY.
-        StreamsGroupMember expectedMember = streamsGroupMemberBuilderWithDefaults(memberId, instanceId)
-                .setMemberEpoch(leaveEpoch)
-                .setPreviousMemberEpoch(memberEpoch - 1)
-                .setTasksPendingRevocation(TasksTupleWithEpochs.EMPTY)
-                .setAssignedTasks(givenAssignedTask)
-                .build();
-
-
-        // WHEN1 : send leave request
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> result = context.streamsGroupHeartbeat(request);
-
-        // THEN1
-        assertResponseEquals(expectedResponseData, result.response().data());
-        assertRecordsEquals(
-                List.of(StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, expectedMember)),
-                result.records()
-        );
-        assertEquals(groupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-
-        // WHEN2 :send leave request again
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> resultFromSecondLeaveMessage = context.streamsGroupHeartbeat(
-                request
-        );
-
-        // THEN2: should return same response.
-        assertResponseEquals(expectedResponseData, resultFromSecondLeaveMessage.response().data());
-        assertRecordsEquals(
-                List.of(StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, expectedMember)),
-                result.records()
-        );
-        assertEquals(groupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-        assertEquals(result, resultFromSecondLeaveMessage);
-    }
-
-    @Test
-    public void testStaticMemberLeaveWithLeaveGroupStaticMemberEpochAndRejoinWithNewMemberId() {
-        String instanceId = Uuid.randomUuid().toString();
-        String oldMemberId = Uuid.randomUuid().toString();
-        String newMemberId = Uuid.randomUuid().toString();
-
-        verifyStaticMemberLeaveAndRejoinNoGroupBump(instanceId, oldMemberId, newMemberId);
-    }
-
-    @Test
-    public void testStaticMemberLeaveWithLeaveGroupStaticMemberEpochAndRejoinWithSameMemberId() {
-        String instanceId = Uuid.randomUuid().toString();
-        String memberId = Uuid.randomUuid().toString();
-
-        verifyStaticMemberLeaveAndRejoinNoGroupBump(instanceId, memberId, memberId);
-    }
-
-    @Test
-    public void testStaticMemberLeaveWithLeaveGroupStaticMemberEpochFromUnrevokedStateAndRejoin() {
-        int leaveEpoch = LEAVE_GROUP_STATIC_MEMBER_EPOCH;
-        int memberEpoch = 10;
-        int groupEpoch = 10;
-
-        String groupId = "fooup";
-        String instanceId = Uuid.randomUuid().toString();
-        String oldMemberId = Uuid.randomUuid().toString();
-        String newMemberId = Uuid.randomUuid().toString();
-
-        String subtopology1 = "subtopology1";
-        String fooTopicName = "foo";
-        Uuid fooTopicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopology1).setSourceTopics(List.of(fooTopicName))
-        ));
-
-        // GIVEN Tasks
-        Map.Entry<String, Set<Integer>> givenAssignedTasks = TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1);
-        Map.Entry<String, Set<Integer>> givenRevokedTasks = TaskAssignmentTestUtil.mkTasks(subtopology1, 2, 3);
-        TasksTupleWithEpochs assignedTasks = mkTasksTupleWithCommonEpoch(
-                TaskRole.ACTIVE, memberEpoch, givenAssignedTasks
-        );
-        TasksTupleWithEpochs tasksPendingRevocation = mkTasksTupleWithCommonEpoch(
-                TaskRole.ACTIVE, memberEpoch, givenRevokedTasks
-        );
-        TasksTuple targetAssignment = TaskAssignmentTestUtil.mkTasksTuple(
-                TaskRole.ACTIVE, givenAssignedTasks
-        );
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(fooTopicId, fooTopicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                fooTopicName, computeTopicHash(fooTopicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(oldMemberId, instanceId)
-                                .setMemberEpoch(memberEpoch)
-                                .setPreviousMemberEpoch(memberEpoch - 1)
-                                .setState(org.apache.kafka.coordinator.group.streams.MemberState.UNREVOKED_TASKS)
-                                .setAssignedTasks(assignedTasks)
-                                .setTasksPendingRevocation(tasksPendingRevocation)
-                                .build())
-                        .withTargetAssignment(oldMemberId, targetAssignment)
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs())
-                )
-                .build();
-
-        // WHEN1 - leave
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> leaveResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(oldMemberId)
-                        .setMemberEpoch(leaveEpoch)
-        );
-
-        // THEN1
-        StreamsGroupHeartbeatResponseData expectedLeaveResponse = new StreamsGroupHeartbeatResponseData()
-                .setMemberId(oldMemberId)
-                .setMemberEpoch(leaveEpoch)
-                .setActiveTasks(null)
-                .setWarmupTasks(null)
-                .setStandbyTasks(null)
-                .setStatus(List.of());
-        assertResponseEquals(expectedLeaveResponse, leaveResult.response().data());
-
-        List<CoordinatorRecord> expectedRecords = List.of(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId,
-                        streamsGroupMemberBuilderWithDefaults(oldMemberId, instanceId)
-                                .setMemberEpoch(leaveEpoch)
-                                .setPreviousMemberEpoch(memberEpoch - 1)
-                                .setAssignedTasks(assignedTasks)
-                                .build())
-        );
-
-        assertRecordsEquals(expectedRecords, leaveResult.records());
-
-        // WHEN2 - re-join
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> rejoinResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(newMemberId)
-                        .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-        );
-
-        // THEN2
-        List<StreamsGroupHeartbeatResponseData.TaskIds> expectedResponseActiveTasks = List.of(
-                new StreamsGroupHeartbeatResponseData.TaskIds()
-                        .setSubtopologyId(subtopology1)
-                        .setPartitions(List.of(0, 1))
-        );
-        StreamsGroupHeartbeatResponseData expectedRejoinResponse = new StreamsGroupHeartbeatResponseData()
-                .setMemberId(newMemberId)
-                .setMemberEpoch(memberEpoch)
-                .setHeartbeatIntervalMs(5000)
-                .setTaskOffsetIntervalMs(60000)
-                .setActiveTasks(expectedResponseActiveTasks)
-                .setWarmupTasks(List.of())
-                .setStandbyTasks(List.of());
-        assertResponseEquals(expectedRejoinResponse, rejoinResult.response().data());
-        assertEquals(org.apache.kafka.coordinator.group.streams.MemberState.STABLE, context.streamsGroupMemberState(groupId, newMemberId));
-        assertEquals(groupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-    }
-
-    @Test
-    public void testStaticMemberLeaveWithLeaveGroupStaticMemberEpochFromUnreleasedStateAndRejoin() {
-        int leaveEpoch = LEAVE_GROUP_STATIC_MEMBER_EPOCH;
-        int memberEpoch = 10;
-        int groupEpoch = 10;
-
-        String groupId = "fooup";
-        String instanceId = Uuid.randomUuid().toString();
-        String oldMemberId = Uuid.randomUuid().toString();
-        String newMemberId = Uuid.randomUuid().toString();
-        String otherMemberId = Uuid.randomUuid().toString();
-        String oldProcessId = Uuid.randomUuid().toString();
-        String otherProcessId = Uuid.randomUuid().toString();
-
-        String subtopology1 = "subtopology1";
-        String fooTopicName = "foo";
-        Uuid fooTopicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopology1).setSourceTopics(List.of(fooTopicName))
-        ));
-
-        // GIVEN Task
-        Map.Entry<String, Set<Integer>> givenAssignedTasks = TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1);
-        Map.Entry<String, Set<Integer>> givenUnreleasedTasks = TaskAssignmentTestUtil.mkTasks(subtopology1, 2);
-        Map.Entry<String, Set<Integer>> givenTargetTasks = TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2);
-        TasksTupleWithEpochs assignedTasks = mkTasksTupleWithCommonEpoch(
-                TaskRole.ACTIVE, memberEpoch, givenAssignedTasks
-        );
-        TasksTupleWithEpochs otherTasksPendingRevocation = mkTasksTupleWithCommonEpoch(
-                TaskRole.ACTIVE, memberEpoch, givenUnreleasedTasks
-        );
-        TasksTuple targetAssignment = TaskAssignmentTestUtil.mkTasksTuple(
-                TaskRole.ACTIVE, givenTargetTasks
-        );
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(fooTopicId, fooTopicName, 3)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                fooTopicName, computeTopicHash(fooTopicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(oldMemberId, instanceId)
-                                .setProcessId(oldProcessId)
-                                .setMemberEpoch(memberEpoch)
-                                .setPreviousMemberEpoch(memberEpoch - 1)
-                                .setState(org.apache.kafka.coordinator.group.streams.MemberState.UNRELEASED_TASKS)
-                                .setAssignedTasks(assignedTasks)
-                                .build())
-                        .withMember(streamsGroupMemberBuilderWithDefaults(otherMemberId)
-                                .setProcessId(otherProcessId)
-                                .setMemberEpoch(memberEpoch)
-                                .setPreviousMemberEpoch(memberEpoch - 1)
-                                .setState(org.apache.kafka.coordinator.group.streams.MemberState.UNREVOKED_TASKS)
-                                .setTasksPendingRevocation(otherTasksPendingRevocation)
-                                .build())
-                        .withTargetAssignment(oldMemberId, targetAssignment)
-                        .withTargetAssignment(otherMemberId, TasksTuple.EMPTY)
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs())
-                )
-                .build();
-
-        // WHEN1 - leave
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> leaveResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(oldMemberId)
-                        .setMemberEpoch(leaveEpoch)
-        );
-
-        // THEN1
-        StreamsGroupHeartbeatResponseData expectedLeaveResponse = new StreamsGroupHeartbeatResponseData()
-                .setMemberId(oldMemberId)
-                .setMemberEpoch(leaveEpoch)
-                .setActiveTasks(null)
-                .setWarmupTasks(null)
-                .setStandbyTasks(null)
-                .setStatus(List.of());
-        assertResponseEquals(expectedLeaveResponse, leaveResult.response().data());
-        List<CoordinatorRecord> expectedRecordTriggeredByLeaving = List.of(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId,
-                        streamsGroupMemberBuilderWithDefaults(oldMemberId, instanceId)
-                                .setProcessId(oldProcessId)
-                                .setMemberEpoch(leaveEpoch)
-                                .setPreviousMemberEpoch(memberEpoch - 1)
-                                .setState(org.apache.kafka.coordinator.group.streams.MemberState.UNRELEASED_TASKS)
-                                .setAssignedTasks(assignedTasks)
-                                .build()));
-        assertRecordsEquals(expectedRecordTriggeredByLeaving, leaveResult.records());
-        assertEquals(org.apache.kafka.coordinator.group.streams.MemberState.UNRELEASED_TASKS, context.streamsGroupMemberState(groupId, oldMemberId));
-
-        // WHEN 2 : rejoin
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> rejoinResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(newMemberId)
-                        .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-        );
-
-        // THEN 2
-        StreamsGroupHeartbeatResponseData expectedRejoinResponse = new StreamsGroupHeartbeatResponseData()
-                .setMemberId(newMemberId)
-                .setMemberEpoch(memberEpoch)
-                .setHeartbeatIntervalMs(5000)
-                .setTaskOffsetIntervalMs(60000)
-                .setActiveTasks(List.of(
-                        new StreamsGroupHeartbeatResponseData.TaskIds()
-                                .setSubtopologyId(subtopology1)
-                                .setPartitions(List.of(0, 1))
-                ))
-                .setWarmupTasks(List.of())
-                .setStandbyTasks(List.of());
-
-        assertResponseEquals(expectedRejoinResponse, rejoinResult.response().data());
-        assertEquals(org.apache.kafka.coordinator.group.streams.MemberState.UNRELEASED_TASKS, context.streamsGroupMemberState(groupId, newMemberId));
-        assertEquals(groupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-    }
-
-    @Test
-    public void testStaticMemberLeaveWithLeaveGroupStaticMemberEpochFromUnrevokedStateAllowsUnreleasedMemberToProgress() {
-        int leaveEpoch = LEAVE_GROUP_STATIC_MEMBER_EPOCH;
-        int memberEpoch = 10;
-        int groupEpoch = 10;
-
-        String groupId = "fooup";
-        String instanceId = Uuid.randomUuid().toString();
-        String leavingMemberId = Uuid.randomUuid().toString();
-        String waitingMemberId = Uuid.randomUuid().toString();
-
-        String subtopology1 = "subtopology1";
-        String fooTopicName = "foo";
-        Uuid fooTopicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopology1).setSourceTopics(List.of(fooTopicName))
-        ));
-
-        // GIVEN Tasks
-        Map.Entry<String, Set<Integer>> leavingAssignedTasks = TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1);
-        Map.Entry<String, Set<Integer>> leavingRevokedTasks = TaskAssignmentTestUtil.mkTasks(subtopology1, 2);
-        TasksTupleWithEpochs assignedTasks = mkTasksTupleWithCommonEpoch(
-                TaskRole.ACTIVE, memberEpoch, leavingAssignedTasks
-        );
-        TasksTupleWithEpochs tasksPendingRevocation = mkTasksTupleWithCommonEpoch(
-                TaskRole.ACTIVE, memberEpoch, leavingRevokedTasks
-        );
-        TasksTuple leavingTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(
-                TaskRole.ACTIVE, leavingAssignedTasks
-        );
-        TasksTuple waitingTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(
-                TaskRole.ACTIVE, leavingRevokedTasks
-        );
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(fooTopicId, fooTopicName, 3)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                fooTopicName, computeTopicHash(fooTopicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(leavingMemberId, instanceId)
-                                .setMemberEpoch(memberEpoch)
-                                .setPreviousMemberEpoch(memberEpoch - 1)
-                                .setState(org.apache.kafka.coordinator.group.streams.MemberState.UNREVOKED_TASKS)
-                                .setAssignedTasks(assignedTasks)
-                                .setTasksPendingRevocation(tasksPendingRevocation)
-                                .build())
-                        .withMember(streamsGroupMemberBuilderWithDefaults(waitingMemberId)
-                                .setMemberEpoch(memberEpoch)
-                                .setPreviousMemberEpoch(memberEpoch)
-                                .setState(org.apache.kafka.coordinator.group.streams.MemberState.UNRELEASED_TASKS)
-                                .build())
-                        .withTargetAssignment(leavingMemberId, leavingTargetAssignment)
-                        .withTargetAssignment(waitingMemberId, waitingTargetAssignment)
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs())
-                )
-                .build();
-
-        // WHEN1 - leave
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> leaveResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(leavingMemberId)
-                        .setMemberEpoch(leaveEpoch)
-        );
-
-        // THEN1
-        StreamsGroupHeartbeatResponseData expectedLeavingResponse = new StreamsGroupHeartbeatResponseData()
-                .setMemberId(leavingMemberId)
-                .setMemberEpoch(leaveEpoch)
-                .setActiveTasks(null)
-                .setWarmupTasks(null)
-                .setStandbyTasks(null)
-                .setStatus(List.of());
-        assertResponseEquals(expectedLeavingResponse, leaveResult.response().data());
-        List<CoordinatorRecord> expectedRecordsTriggeredByLeave = List.of(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId,
-                        streamsGroupMemberBuilderWithDefaults(leavingMemberId, instanceId)
-                                .setMemberEpoch(leaveEpoch)
-                                .setPreviousMemberEpoch(memberEpoch - 1)
-                                .setAssignedTasks(assignedTasks)
-                                .build()));
-        assertRecordsEquals(expectedRecordsTriggeredByLeave, leaveResult.records());
-        assertEquals(org.apache.kafka.coordinator.group.streams.MemberState.STABLE, context.streamsGroupMemberState(groupId, leavingMemberId));
-
-        // When2 - Waiting member send a heartbeat expecting get unreleased tasks.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> waitingMemberResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setMemberId(waitingMemberId)
-                        .setMemberEpoch(memberEpoch)
-        );
-
-        // THEN2
-        StreamsGroupHeartbeatResponseData expectedWaitngMemberResponse = new StreamsGroupHeartbeatResponseData()
-                .setMemberId(waitingMemberId)
-                .setMemberEpoch(memberEpoch)
-                .setHeartbeatIntervalMs(5000)
-                .setTaskOffsetIntervalMs(60000)
-                .setActiveTasks(List.of(
-                        new StreamsGroupHeartbeatResponseData.TaskIds()
-                                .setSubtopologyId(subtopology1)
-                                .setPartitions(List.of(2))
-                ))
-                .setWarmupTasks(List.of())
-                .setStandbyTasks(List.of());
-        assertResponseEquals(expectedWaitngMemberResponse, waitingMemberResult.response().data());
-
-
-        List<CoordinatorRecord> expectedRecordsTriggeredByWaitngMember = List.of(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId,
-                        streamsGroupMemberBuilderWithDefaults(waitingMemberId)
-                                .setMemberEpoch(memberEpoch)
-                                .setPreviousMemberEpoch(memberEpoch)
-                                .setAssignedTasks(mkTasksTupleWithCommonEpoch(
-                                        TaskRole.ACTIVE, memberEpoch, leavingRevokedTasks
-                                ))
-                                .build()));
-        assertRecordsEquals(expectedRecordsTriggeredByWaitngMember, waitingMemberResult.records());
-
-        assertEquals(org.apache.kafka.coordinator.group.streams.MemberState.STABLE, context.streamsGroupMemberState(groupId, waitingMemberId));
-        assertEquals(groupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-    }
-
-    private void verifyStaticMemberLeaveAndRejoinNoGroupBump(String instanceId, String oldMemberId, String newMemberId) {
-        /*
-         * Verifies:
-         * 1. leave(-2) does not bump group epoch.
-         * 2. rejoin restores member epoch and assignment.
-         * 3. replacement/tombstone records are written as expected.
-         */
-
-        // GIVEN
-        int leaveEpoch = LEAVE_GROUP_STATIC_MEMBER_EPOCH;
-        int memberEpoch = 10;
-        int groupEpoch = 10;
-
-        String groupId = "fooup";
-
-        String subtopology1 = "subtopology1";
-        String fooTopicName = "foo";
-        Uuid fooTopicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopology1).setSourceTopics(List.of(fooTopicName))
-        ));
-
-        // GIVEN Task
-        Map.Entry<String, Set<Integer>> givenTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3);
-        TasksTupleWithEpochs givenAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, 10, givenTaskEntries);
-        TasksTuple givenTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, givenTaskEntries);
-        List<StreamsGroupHeartbeatResponseData.TaskIds> taskIds = List.of(
-                new StreamsGroupHeartbeatResponseData.TaskIds()
-                        .setSubtopologyId(subtopology1)
-                        .setPartitions(List.of(0, 1, 2, 3)
-                        )
-        );
-
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(fooTopicId, fooTopicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                fooTopicName, computeTopicHash(fooTopicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(oldMemberId, instanceId)
-                                .setMemberEpoch(memberEpoch)
-                                .setPreviousMemberEpoch(memberEpoch - 1)
-                                .setAssignedTasks(givenAssignedTasks)
-                                .build())
-                        .withTargetAssignment(oldMemberId, givenTargetAssignment)
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs())
-                )
-                .build();
-
-        // WHEN1 : normal heart beat.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> normalHeartbeatResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(oldMemberId)
-                        .setMemberEpoch(memberEpoch)
-        );
-
-        // THEN1 : 
-        // - all tasks should be null because assigned tasks unchanged.
-        // - Keep the group epoch.
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberEpoch(memberEpoch)
-                        .setMemberId(oldMemberId)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(null)
-                        .setWarmupTasks(null)
-                        .setStandbyTasks(null),
-                normalHeartbeatResult.response().data()
-        );
-        assertEquals(groupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-
-
-        // WHEN2 : Stream Member leave with -2
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> leaveResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(oldMemberId)
-                        .setMemberEpoch(leaveEpoch)
-        );
-
-        // THEN2
-        // - Keep the group epoch.
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(oldMemberId)
-                        .setMemberEpoch(leaveEpoch)
-                        .setActiveTasks(null)
-                        .setWarmupTasks(null)
-                        .setStandbyTasks(null)
-                        .setStatus(List.of()),
-                leaveResult.response().data()
-        );
-        assertEquals(groupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-
-        // WHEN3 : Streams Member rejoin with other memberId
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> rejoinResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(newMemberId)
-                        .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-        );
-
-        // THEN3 : 
-        // - Inherit previous member's member epoch, and assigned tasks.
-        // - Keep the member epoch bump.
-        // - Keep the group epoch bump.
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(newMemberId)
-                        .setMemberEpoch(memberEpoch)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(taskIds)
-                        .setWarmupTasks(List.of())
-                        .setStandbyTasks(List.of()),
-                rejoinResult.response().data()
-        );
-        assertEquals(groupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-
-        StreamsGroupMember newJoinStaticMember = streamsGroupMemberBuilderWithDefaults(newMemberId, instanceId)
-                .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-                .setPreviousMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-                .setAssignedTasks(givenAssignedTasks)
-                .build();
-
-        StreamsGroupMember withPrevMemberId = streamsGroupMemberBuilderWithDefaults(newMemberId, instanceId)
-                .setMemberEpoch(memberEpoch) // 0 -> 10
-                .setPreviousMemberEpoch(0) //  0 -> 0
-                .setAssignedTasks(givenAssignedTasks)
-                .build();
-
-        assertRecordsEquals(
-                List.of(
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentTombstoneRecord(groupId, oldMemberId),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentTombstoneRecord(groupId, oldMemberId),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupMemberTombstoneRecord(groupId, oldMemberId),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupMemberRecord(groupId, newJoinStaticMember),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentRecord(groupId, newJoinStaticMember.memberId(), givenTargetAssignment),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, newJoinStaticMember),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, withPrevMemberId)
-                ),
-                rejoinResult.records()
-        );
-    }
-
-    @Test
-    public void testStaticMemberLeaveWithLeaveGroupStaticMemberEpochAndRejoinAndOtherRackIdThenGroupBumpOccur() {
-        // GIVEN
-        int leaveEpoch = LEAVE_GROUP_STATIC_MEMBER_EPOCH;
-        int memberEpoch = 10;
-        int groupEpoch = 10;
-
-        String groupId = "fooup";
-        String rackId = Uuid.randomUuid().toString();
-        String memberId = Uuid.randomUuid().toString();
-        String instanceId = Uuid.randomUuid().toString();
-
-        String subtopology1 = "subtopology1";
-        String fooTopicName = "foo";
-        Uuid fooTopicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopology1).setSourceTopics(List.of(fooTopicName))
-        ));
-
-        // GIVEN Task
-        Map.Entry<String, Set<Integer>> givenTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3);
-        TasksTupleWithEpochs givenAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, 10, givenTaskEntries);
-        TasksTuple givenTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, givenTaskEntries);
-        List<StreamsGroupHeartbeatResponseData.TaskIds> taskIds = List.of(
-                new StreamsGroupHeartbeatResponseData.TaskIds()
-                        .setSubtopologyId(subtopology1)
-                        .setPartitions(List.of(0, 1, 2, 3)
-                        )
-        );
-
-        // GIVEN Assignor
-        MockTaskAssignor assignor = new MockTaskAssignor("sticky");
-
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(fooTopicId, fooTopicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                fooTopicName, computeTopicHash(fooTopicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(assignor))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(memberId, instanceId)
-                                .setMemberEpoch(memberEpoch)
-                                .setRackId(rackId)
-                                .setPreviousMemberEpoch(memberEpoch - 1)
-                                .setAssignedTasks(givenAssignedTasks)
-                                .build())
-                        .withTargetAssignment(memberId, givenTargetAssignment)
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs())
-                )
-                .build();
-
-        // WHEN1 : normal heart beat.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> normalHeartbeatResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setRackId(rackId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(memberId)
-                        .setMemberEpoch(memberEpoch)
-        );
-
-        // THEN1 : 
-        // - all tasks should be null because assigned tasks unchanged.
-        // - Keep the group epoch.
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberEpoch(memberEpoch)
-                        .setMemberId(memberId)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(null)
-                        .setWarmupTasks(null)
-                        .setStandbyTasks(null),
-                normalHeartbeatResult.response().data()
-        );
-        assertEquals(groupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-
-
-        // WHEN2 : Stream Member leave with -2
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> leaveResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setRackId(rackId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(memberId)
-                        .setMemberEpoch(leaveEpoch)
-        );
-
-        // THEN2
-        // - Keep the group epoch.
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(memberId)
-                        .setMemberEpoch(leaveEpoch)
-                        .setActiveTasks(null)
-                        .setWarmupTasks(null)
-                        .setStandbyTasks(null)
-                        .setStatus(List.of()),
-                leaveResult.response().data()
-        );
-        assertEquals(groupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-
-        // GIVEN3
-        String newMemberId = Uuid.randomUuid().toString();
-        String newRackId = Uuid.randomUuid().toString();
-        assignor.prepareGroupAssignment(Map.of(newMemberId, givenTargetAssignment));
-
-        int bumpedGroupEpoch = groupEpoch + 1;
-        int bumpedMemberEpoch = memberEpoch + 1;
-
-        // WHEN3 : Streams Member rejoin with other memberId and rackId
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> rejoinResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setRackId(newRackId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(newMemberId)
-                        .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-        );
-
-        // THEN3 : 
-        // - Inherit previous member's member epoch, and assigned tasks.
-        // - member epoch should be bumped.
-        // - group epoch should be bumped.
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(newMemberId)
-                        .setMemberEpoch(bumpedMemberEpoch)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(taskIds)
-                        .setWarmupTasks(List.of())
-                        .setStandbyTasks(List.of()),
-                rejoinResult.response().data()
-        );
-        assertEquals(bumpedGroupEpoch, context.groupMetadataManager.streamsGroup(groupId).groupEpoch());
-
-        StreamsGroupMember transationStaticInitMember = streamsGroupMemberBuilderWithDefaults(newMemberId, instanceId)
-                .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-                .setPreviousMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-                .setRackId(rackId)
-                .setAssignedTasks(givenAssignedTasks)
-                .build();
-
-        StreamsGroupMember newJoinStaticMember = streamsGroupMemberBuilderWithDefaults(newMemberId, instanceId)
-                .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-                .setPreviousMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-                .setRackId(newRackId)
-                .setAssignedTasks(givenAssignedTasks)
-                .build();
-
-        StreamsGroupMember reconciledMember = streamsGroupMemberBuilderWithDefaults(newMemberId, instanceId)
-                .setMemberEpoch(bumpedMemberEpoch)
-                .setPreviousMemberEpoch(0)
-                .setRackId(newRackId)
-                .setAssignedTasks(givenAssignedTasks)
-                .build();
-
-        assertRecordsEquals(
-                List.of(
-                        // From eplaceStreamsMembers
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentTombstoneRecord(groupId, memberId),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentTombstoneRecord(groupId, memberId),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupMemberTombstoneRecord(groupId, memberId),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupMemberRecord(groupId, transationStaticInitMember),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentRecord(groupId, transationStaticInitMember.memberId(), givenTargetAssignment),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, transationStaticInitMember),
-
-                        // From hasStreamsMemberMetadataChanged
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupMemberRecord(groupId, newJoinStaticMember),
-
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupMetadataRecord(groupId, bumpedGroupEpoch, groupMetadataHash, 0, getDefaultAssignmentConfigs()),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentRecord(groupId, newJoinStaticMember.memberId(), givenTargetAssignment),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentMetadataRecord(groupId, bumpedGroupEpoch, context.time.milliseconds()),
-                        StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, reconciledMember)
-                ),
-                rejoinResult.records()
-        );
-    }
-
-
-    @Test
-    public void testStaticMemberRejoinWritesReplacementRecordsInStreamsGroup() {
-        String groupId = "fooup";
-        String oldMemberId = Uuid.randomUuid().toString();
-        String rejoinMemberId = Uuid.randomUuid().toString();
-        String instanceId = Uuid.randomUuid().toString();
-
-        String subtopology1 = "subtopology1";
-        String fooTopicName = "foo";
-        Uuid fooTopicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopology1).setSourceTopics(List.of(fooTopicName))
-        ));
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(fooTopicId, fooTopicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                fooTopicName, computeTopicHash(fooTopicName, metadataImage)
-        ));
-        TasksTuple oldTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE,
-                TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3));
-
-        StreamsGroupMember oldMember = streamsGroupMemberBuilderWithDefaults(oldMemberId, instanceId)
-                .setMemberEpoch(-2)
-                .setPreviousMemberEpoch(10)
-                .setAssignedTasks(mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, 10,
-                        TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3)))
-                .build();
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, 10)
-                        .withMember(oldMember)
-                        .withTargetAssignment(oldMemberId, oldTargetAssignment)
-                        .withTargetAssignmentEpoch(10)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs())
-                )
-                .build();
-
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> result = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId)
-                        .setMemberId(rejoinMemberId)
-                        .setMemberEpoch(0)
-                        .setProcessId(DEFAULT_PROCESS_ID)
-                        .setActiveTasks(List.of())
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of())
-        );
-
-        assertEquals(rejoinMemberId, result.response().data().memberId());
-        assertEquals(10, result.response().data().memberEpoch());
-
-        StreamsGroupMember expectedCopiedMember = new StreamsGroupMember.Builder(oldMember, rejoinMemberId)
-                .setMemberEpoch(0)
-                .setPreviousMemberEpoch(0)
-                .build();
-
-        assertTrue(result.records().contains(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentTombstoneRecord(groupId, oldMemberId)
-        ));
-        assertTrue(result.records().contains(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentTombstoneRecord(groupId, oldMemberId)
-        ));
-        assertTrue(result.records().contains(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMemberTombstoneRecord(groupId, oldMemberId)
-        ));
-        assertTrue(result.records().contains(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMemberRecord(groupId, expectedCopiedMember)
-        ));
-        assertTrue(result.records().contains(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentRecord(groupId, rejoinMemberId, oldTargetAssignment)
-        ));
-        assertTrue(result.records().contains(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, expectedCopiedMember)
-        ));
-    }
-
-    @Test
-    public void testStaticMemberLeaveWithMismatchedMemberIdThrowsFencedInstanceIdInStreamsGroup() {
-        String groupId = "fooup";
-        String memberId = Uuid.randomUuid().toString();
-        String differentMemberId = Uuid.randomUuid().toString();
-        String instanceId = Uuid.randomUuid().toString();
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, 10)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(memberId, instanceId)
-                                .setMemberEpoch(10)
-                                .setPreviousMemberEpoch(9)
-                                .build())
-                        .withTargetAssignmentEpoch(10)
-                )
-                .build();
-
-        assertThrows(FencedInstanceIdException.class, () ->
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setInstanceId(instanceId)
-                                .setMemberId(differentMemberId)
-                                .setMemberEpoch(-2)
-                ));
-    }
-
-    @Test
-    public void testUnknownStaticMemberHeartbeatWithPositiveEpochThrowsUnknownMemberIdInStreamsGroup() {
-        String groupId = "fooup";
-        String memberId = Uuid.randomUuid().toString();
-        String instanceId = Uuid.randomUuid().toString();
-        String unknownInstanceId = Uuid.randomUuid().toString();
-        String unknownMemberId = Uuid.randomUuid().toString();
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, 10)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(memberId, instanceId)
-                                .setMemberEpoch(10)
-                                .setPreviousMemberEpoch(9)
-                                .build())
-                        .withTargetAssignmentEpoch(10)
-                )
-                .build();
-
-        UnknownMemberIdException e = assertThrows(UnknownMemberIdException.class, () ->
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setInstanceId(unknownInstanceId)
-                                .setMemberId(unknownMemberId)
-                                .setMemberEpoch(1)
-                ));
-        assertEquals(String.format("Instance id %s is unknown.", unknownInstanceId), e.getMessage());
-    }
-
-    private void testStaticMemberJoinThenRevokeAndReceiveTasksWith2Members(int maxSize) {
-        String groupId = "fooup";
-
-        String memberId1 = Uuid.randomUuid().toString();
-        String memberId2 = Uuid.randomUuid().toString();
-        String otherMemberId2 = Uuid.randomUuid().toString();
-
-        String instanceId1 = Uuid.randomUuid().toString();
-        String instanceId2 = Uuid.randomUuid().toString();
-
-        String subtopology1 = "subtopology1";
-        String fooTopicName = "foo";
-        Uuid fooTopicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopology1).setSourceTopics(List.of(fooTopicName))
-        ));
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(fooTopicId, fooTopicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                fooTopicName, computeTopicHash(fooTopicName, metadataImage)
-        ));
-
-        MockTaskAssignor assignor = new MockTaskAssignor("sticky");
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withConfig(GroupCoordinatorConfig.STREAMS_GROUP_MAX_SIZE_CONFIG, maxSize)
-                .withStreamsGroupTaskAssignors(List.of(assignor))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, 10)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(memberId1)
-                                .setInstanceId(instanceId1)
-                                .setMemberEpoch(10)
-                                .setPreviousMemberEpoch(9)
-                                .setAssignedTasks(mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, 10,
-                                        TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3)))
-                                .build())
-                        .withTargetAssignment(memberId1, TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE,
-                                TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1, 2, 3)))
-                        .withTargetAssignmentEpoch(10)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withMetadataHash(groupMetadataHash)
-                        .withValidatedTopologyEpoch(0)
-                )
-                .build();
-
-        // Next target assignment after member2 joins.
-        assignor.prepareGroupAssignment(Map.of(
-                memberId1, TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE,
-                        TaskAssignmentTestUtil.mkTasks(subtopology1, 0, 1)
-                ),
-                memberId2, TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE,
-                        TaskAssignmentTestUtil.mkTasks(subtopology1, 2, 3)
-                )
-        ));
-
-        // 1) Static member2 joins. It gets no active tasks yet because member1 still owns them.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> joinResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId2)
-                        .setMemberId(memberId2)
-                        .setMemberEpoch(0)
-                        .setRebalanceTimeoutMs(1500)
-                        .setTopology(topology)
-                        .setProcessId(DEFAULT_PROCESS_ID)
-                        .setActiveTasks(List.of())
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of())
-        );
-
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(memberId2)
-                        .setMemberEpoch(11)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(List.of())
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of())
-                        .setStatus(List.of()),
-                joinResult.response().data()
-        );
-
-        // 2) member1 receives revocation instruction: keep only [0,1].
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> revokeInstructionResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId1)
-                        .setMemberId(memberId1)
-                        .setMemberEpoch(10)
-        );
-
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(memberId1)
-                        .setMemberEpoch(10)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(List.of(
-                                new StreamsGroupHeartbeatResponseData.TaskIds()
-                                        .setSubtopologyId(subtopology1)
-                                        .setPartitions(List.of(0, 1))
-                        ))
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of())
-                        .setStatus(List.of()),
-                revokeInstructionResult.response().data()
-        );
-
-        // 3) member1 acknowledges revocation by reporting owned active tasks [0,1].
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> revokeAckResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId1)
-                        .setMemberId(memberId1)
-                        .setMemberEpoch(10)
-                        .setActiveTasks(List.of(
-                                new StreamsGroupHeartbeatRequestData.TaskIds()
-                                        .setSubtopologyId(subtopology1)
-                                        .setPartitions(List.of(0, 1))
-                        ))
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of())
-        );
-
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(memberId1)
-                        .setMemberEpoch(11)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setStatus(List.of()),
-                revokeAckResult.response().data()
-        );
-
-        // 4) member2 heartbeats again and now receives [2,3].
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> member2ReceiveResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId2)
-                        .setMemberId(memberId2)
-                        .setMemberEpoch(11)
-        );
-
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(memberId2)
-                        .setMemberEpoch(11)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(List.of(
-                                new StreamsGroupHeartbeatResponseData.TaskIds()
-                                        .setSubtopologyId(subtopology1)
-                                        .setPartitions(List.of(2, 3))
-                        ))
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of())
-                        .setStatus(List.of()),
-                member2ReceiveResult.response().data()
-        );
-
-        // 5) member2 leave.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> member2LeaveResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId2)
-                        .setMemberId(memberId2)
-                        .setMemberEpoch(-2)
-        );
-
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(memberId2)
-                        .setMemberEpoch(-2)
-                        .setHeartbeatIntervalMs(0)
-                        .setActiveTasks(null)
-                        .setStandbyTasks(null)
-                        .setWarmupTasks(null)
-                        .setStatus(List.of()),
-                member2LeaveResult.response().data()
-        );
-
-        // 6) member2 re-join with other memberId.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> member2rejoinResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(instanceId2)
-                        .setMemberId(otherMemberId2)
-                        .setMemberEpoch(0)
-        );
-
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(otherMemberId2)
-                        .setMemberEpoch(11)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(List.of(
-                                new StreamsGroupHeartbeatResponseData.TaskIds()
-                                        .setSubtopologyId(subtopology1)
-                                        .setPartitions(List.of(2, 3))
-                        ))
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of())
-                        .setStatus(List.of()),
-                member2rejoinResult.response().data()
-        );
-    }
-
-    @Test
-    public void testDynamicJoinFailsAtMaxSizeWhileStaticMemberIsTemporarilyLeftAndDynamicMemberStillExists() {
-        // STREAMS_GROUP_MAX_SIZE_CONFIG is 2. 
-        // There are 2 members. (1 dynamic member, 1 static member)
-        //   - one static member leaves temporarily with -2 epoch.
-        //   - one dynamic member alive in group.
-        // Another dynamic member try to join.
-        
-        String groupId = "fooup";
-        String staticMemberId = Uuid.randomUuid().toString();
-        String staticInstanceId = Uuid.randomUuid().toString();
-        String dynamicMemberId = Uuid.randomUuid().toString();
-        String newDynamicMemberId = Uuid.randomUuid().toString();
-        Topology topology = new Topology().setSubtopologies(List.of());
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withMetadataImage(new MetadataImageBuilder().buildCoordinatorMetadataImage())
-                .withConfig(GroupCoordinatorConfig.STREAMS_GROUP_MAX_SIZE_CONFIG, 2)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, 10)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(staticMemberId, staticInstanceId)
-                                .setMemberEpoch(StreamsGroupHeartbeatRequest.LEAVE_GROUP_STATIC_MEMBER_EPOCH)
-                                .setPreviousMemberEpoch(10)
-                                .build())
-                        .withMember(streamsGroupMemberBuilderWithDefaults(dynamicMemberId)
-                                .setMemberEpoch(10)
-                                .setPreviousMemberEpoch(9)
-                                .build())
-                        .withTargetAssignmentEpoch(10)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology)))
-                .build();
-
-        assertThrows(GroupMaxSizeReachedException.class, () ->
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setMemberId(newDynamicMemberId)
-                                .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-                                .setProcessId("new-process-id")
-                                .setRebalanceTimeoutMs(1500)
-                                .setTopology(topology)
-                                .setActiveTasks(List.of())
-                                .setStandbyTasks(List.of())
-                                .setWarmupTasks(List.of())
-                )
-        );
-    }
-
-    @Test
-    public void testStaticRejoinSucceedsAtMaxSizeWhileDynamicMemberStillExists() {
-        // Scenario
-        // STREAMS_GROUP_MAX_SIZE_CONFIG is 2. 
-        // There are 2 members. (1 dynamic member, 1 static member)
-        //   - static member leaves temporarily with -2 epoch.
-        //   - dynamic member alive in group
-        // static member try to rejoin.
-        
-        int groupEpoch = 10;
-        String groupId = "fooup";
-
-        String subtopologyId = "subtopology-1";
-        String topicName = "foo";
-        Uuid topicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopologyId).setSourceTopics(List.of(topicName))
-        ));
-
-        String oldStaticMemberId = Uuid.randomUuid().toString();
-        String newStaticMemberId = Uuid.randomUuid().toString();
-        String staticInstanceId = Uuid.randomUuid().toString();
-        String dynamicMemberId = Uuid.randomUuid().toString();
-
-        // GIVEN Task for static member
-        Map.Entry<String, Set<Integer>> staticTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId, 0, 1);
-        TasksTupleWithEpochs staticAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, staticTaskEntries);
-        TasksTuple staticTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, staticTaskEntries);
-        
-        // GIVEN Task for dynamic member
-        Map.Entry<String, Set<Integer>> dynamicTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId,2, 3);
-        TasksTupleWithEpochs dynamicAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, dynamicTaskEntries);
-        TasksTuple dynamicTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, dynamicTaskEntries);
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(topicId, topicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                topicName, computeTopicHash(topicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withConfig(GroupCoordinatorConfig.STREAMS_GROUP_MAX_SIZE_CONFIG, 2)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(oldStaticMemberId, staticInstanceId)
-                                .setMemberEpoch(StreamsGroupHeartbeatRequest.LEAVE_GROUP_STATIC_MEMBER_EPOCH)
-                                .setPreviousMemberEpoch(10)
-                                .setAssignedTasks(staticAssignedTasks)
-                                .build())
-                        .withMember(streamsGroupMemberBuilderWithDefaults(dynamicMemberId)
-                                .setMemberEpoch(10)
-                                .setPreviousMemberEpoch(9)
-                                .setAssignedTasks(dynamicAssignedTasks)
-                                .build())
-                        .withTargetAssignment(oldStaticMemberId, staticTargetAssignment)
-                        .withTargetAssignment(dynamicMemberId, dynamicTargetAssignment)
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs()))
-                .build();
-
-        // WHEN - static member rejoin.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> rejoinResult = context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(staticInstanceId)
-                        .setMemberId(newStaticMemberId)
-                        .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-        );
-
-        // THEN - At the maxsize, static member still can rejoin if static member leaves with epoch -2.
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(newStaticMemberId)
-                        .setMemberEpoch(groupEpoch)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(mkResponseTasks(subtopologyId, 0, 1))
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of()),
-                rejoinResult.response().data()
-        );
-
-        StreamsGroup group = context.groupMetadataManager.streamsGroup(groupId);
-        assertFalse(group.hasMember(oldStaticMemberId));
-        assertTrue(group.hasMember(newStaticMemberId));
-        assertEquals(newStaticMemberId, group.staticMember(staticInstanceId).memberId());
-        
-        assertTrue(group.hasMember(dynamicMemberId));
-        assertEquals(dynamicAssignedTasks, group.getMemberOrThrow(dynamicMemberId).assignedTasks());
-        assertEquals(dynamicTargetAssignment, group.targetAssignment(dynamicMemberId, Optional.empty()));
-        assertEquals(groupEpoch, group.groupEpoch());
-    }
-
-
-    @Test
-    public void testStaticTemporaryLeaveDoesNotTransferTasksToExistingDynamicMember() {
-        // Scenario
-        // STREAMS_GROUP_MAX_SIZE_CONFIG is 2. 
-        // There are 2 members. (1 dynamic member, 1 static member)
-        //   - static member leaves temporarily with -2 epoch.
-        //   - dynamic member alive in group
-        // If dynamic send a heartbeat, there is no new assignment. 
-        int groupEpoch = 10;
-        String groupId = "fooup";
-
-        String subtopologyId = "subtopology-1";
-        String topicName = "foo";
-        Uuid topicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopologyId).setSourceTopics(List.of(topicName))
-        ));
-
-        String staticMemberId = Uuid.randomUuid().toString();
-        String staticInstanceId = Uuid.randomUuid().toString();
-        String dynamicMemberId = Uuid.randomUuid().toString();
-
-        // GIVEN Task for static member
-        Map.Entry<String, Set<Integer>> staticTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId,0, 1);
-        TasksTupleWithEpochs staticAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, staticTaskEntries);
-        TasksTuple staticTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, staticTaskEntries);
-        
-        // GIVEN Task for dynamic member
-        Map.Entry<String, Set<Integer>> dynamicTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId, 2);
-        TasksTupleWithEpochs dynamicAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, dynamicTaskEntries);
-        TasksTuple dynamicTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, dynamicTaskEntries);
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(topicId, topicName, 3)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                topicName, computeTopicHash(topicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(staticMemberId, staticInstanceId)
-                                .setMemberEpoch(groupEpoch)
-                                .setPreviousMemberEpoch(groupEpoch - 1)
-                                .setAssignedTasks(staticAssignedTasks)
-                                .build())
-                        .withMember(streamsGroupMemberBuilderWithDefaults(dynamicMemberId)
-                                .setMemberEpoch(groupEpoch)
-                                .setPreviousMemberEpoch(groupEpoch - 1)
-                                .setAssignedTasks(dynamicAssignedTasks)
-                                .build())
-                        .withTargetAssignment(staticMemberId, staticTargetAssignment)
-                        .withTargetAssignment(dynamicMemberId, dynamicTargetAssignment)
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs()))
-                .build();
-
-        // WHEN - static member leaves with epoch -2.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> leaveResult =
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setInstanceId(staticInstanceId)
-                                .setMemberId(staticMemberId)
-                                .setMemberEpoch(StreamsGroupHeartbeatRequest.LEAVE_GROUP_STATIC_MEMBER_EPOCH)
-                );
-
-        // THEN
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(staticMemberId)
-                        .setMemberEpoch(StreamsGroupHeartbeatRequest.LEAVE_GROUP_STATIC_MEMBER_EPOCH)
-                        .setActiveTasks(null)
-                        .setWarmupTasks(null)
-                        .setStandbyTasks(null)
-                        .setStatus(List.of()),
-                leaveResult.response().data()
-        );
-
-        // WHEN2 - dynamic member send a heartbeat.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> dynamicHeartbeatResult =
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setMemberId(dynamicMemberId)
-                                .setMemberEpoch(groupEpoch)
-                );
-
-        // THEN2 : There is no new assignment.
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(dynamicMemberId)
-                        .setMemberEpoch(groupEpoch)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(null)
-                        .setWarmupTasks(null)
-                        .setStandbyTasks(null),
-                dynamicHeartbeatResult.response().data()
-        );
-        assertTrue(dynamicHeartbeatResult.records().isEmpty());
-
-        StreamsGroup group = context.groupMetadataManager.streamsGroup(groupId);
-        assertEquals(dynamicAssignedTasks, group.getMemberOrThrow(dynamicMemberId).assignedTasks());
-        assertEquals(dynamicTargetAssignment, group.targetAssignment(dynamicMemberId, Optional.empty()));
-        assertEquals(groupEpoch, group.groupEpoch());
-    }
-
-    @Test
-    public void testStaticRejoinWithUpdatedProcessIdRecomputesTargetAssignmentAndDynamicMemberReconcilesInMixedGroup() {
-        // Scenario:
-        // There are 2 members.
-        //   - static member left temporarily with -2 epoch.
-        //   - dynamic member alive in group.
-        // When the same static instance rejoins with a different processId, the coordinator
-        // bumps the group epoch and recomputes the target assignment.
-        // The dynamic member keeps its current assignment until its next heartbeat.
-        // When the dynamic member sends the next heartbeat, it reconciles to the recomputed target assignment.
-
-        String groupId = "fooup";
-        int groupEpoch = 10;
-        int bumpedGroupEpoch = groupEpoch + 1;
-
-        String subtopologyId = "subtopology-1";
-        String topicName = "foo";
-        Uuid topicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopologyId).setSourceTopics(List.of(topicName))
-        ));
-
-        String oldStaticMemberId = Uuid.randomUuid().toString();
-        String rejoinStaticMemberId = Uuid.randomUuid().toString();
-        String staticInstanceId = Uuid.randomUuid().toString();
-        String dynamicMemberId = Uuid.randomUuid().toString();
-
-        String oldProcessId = "old-process-id";
-        String newProcessId = "new-process-id";
-
-        // Initial assignment
-        Map.Entry<String, Set<Integer>> staticTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId, 0, 1);
-        TasksTupleWithEpochs staticAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, staticTaskEntries);
-        TasksTuple oldStaticTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, staticTaskEntries);
-
-        Map.Entry<String, Set<Integer>> dynamicTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId, 2, 3);
-        TasksTupleWithEpochs dynamicAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, dynamicTaskEntries);
-        TasksTuple oldDynamicTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, dynamicTaskEntries);
-
-        // Recomputed target assignment after static member rejoins with updated processId
-        TasksTuple newStaticTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(
-                TaskRole.ACTIVE,
-                TaskAssignmentTestUtil.mkTasks(subtopologyId, 0)
-        );
-        TasksTuple newDynamicTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(
-                TaskRole.ACTIVE,
-                TaskAssignmentTestUtil.mkTasks(subtopologyId, 1, 2, 3)
-        );
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(topicId, topicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                topicName, computeTopicHash(topicName, metadataImage)
-        ));
-
-        MockTaskAssignor assignor = new MockTaskAssignor("sticky");
-        assignor.prepareGroupAssignment(Map.of(
-                rejoinStaticMemberId, newStaticTargetAssignment,
-                dynamicMemberId, newDynamicTargetAssignment
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(assignor))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(oldStaticMemberId, staticInstanceId)
-                                .setMemberEpoch(StreamsGroupHeartbeatRequest.LEAVE_GROUP_STATIC_MEMBER_EPOCH)
-                                .setPreviousMemberEpoch(groupEpoch)
-                                .setProcessId(oldProcessId)
-                                .setAssignedTasks(staticAssignedTasks)
-                                .build())
-                        .withMember(streamsGroupMemberBuilderWithDefaults(dynamicMemberId)
-                                .setMemberEpoch(groupEpoch)
-                                .setPreviousMemberEpoch(groupEpoch - 1)
-                                .setAssignedTasks(dynamicAssignedTasks)
-                                .build())
-                        .withTargetAssignment(oldStaticMemberId, oldStaticTargetAssignment)
-                        .withTargetAssignment(dynamicMemberId, oldDynamicTargetAssignment)
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs()))
-                .build();
-
-        // WHEN 1: static member try to rejoin with new process id.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> rejoinResult =
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setInstanceId(staticInstanceId)
-                                .setMemberId(rejoinStaticMemberId)
-                                .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-                                .setProcessId(newProcessId)
-                                .setActiveTasks(List.of())
-                                .setStandbyTasks(List.of())
-                                .setWarmupTasks(List.of())
-                );
-
-        // THEN 1:
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(rejoinStaticMemberId)
-                        .setMemberEpoch(bumpedGroupEpoch)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(mkResponseTasks(subtopologyId, 0))
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of()),
-                rejoinResult.response().data()
-        );
-
-        StreamsGroup group = context.groupMetadataManager.streamsGroup(groupId);
-
-        // group epoch should be bumped up.
-        assertEquals(bumpedGroupEpoch, group.groupEpoch());
-        
-        // static member should be replaced.
-        assertFalse(group.hasMember(oldStaticMemberId));
-        assertTrue(group.hasMember(rejoinStaticMemberId));
-        assertEquals(rejoinStaticMemberId, group.staticMember(staticInstanceId).memberId());
-        
-        // Because group epoch is bumped up, target assignment should be recomupted.
-        assertEquals(newStaticTargetAssignment, group.targetAssignment(rejoinStaticMemberId, Optional.of(staticInstanceId)));
-        assertEquals(newDynamicTargetAssignment, group.targetAssignment(dynamicMemberId, Optional.empty()));
-        assertEquals(dynamicAssignedTasks, group.getMemberOrThrow(dynamicMemberId).assignedTasks());
-        assertEquals(groupEpoch, group.getMemberOrThrow(dynamicMemberId).memberEpoch());
-
-        StreamsGroupMember expectedCopiedStaticMember = streamsGroupMemberBuilderWithDefaults(rejoinStaticMemberId, staticInstanceId)
-                .setProcessId(oldProcessId)
-                .setMemberEpoch(0)
-                .setPreviousMemberEpoch(0)
-                .setAssignedTasks(staticAssignedTasks)
-                .build();
-
-        StreamsGroupMember expectedUpdatedStaticMember = streamsGroupMemberBuilderWithDefaults(rejoinStaticMemberId, staticInstanceId)
-                .setProcessId(newProcessId)
-                .setMemberEpoch(0)
-                .setPreviousMemberEpoch(0)
-                .setAssignedTasks(staticAssignedTasks)
-                .build();
-
-        StreamsGroupMember expectedReconciledStaticMember = streamsGroupMemberBuilderWithDefaults(rejoinStaticMemberId, staticInstanceId)
-                .setProcessId(newProcessId)
-                .setMemberEpoch(bumpedGroupEpoch)
-                .setPreviousMemberEpoch(0)
-                .setAssignedTasks(mkTasksTupleWithEpochs(
-                        TaskRole.ACTIVE,
-                        mkTasksWithEpochs(subtopologyId, Map.of(0, groupEpoch))
-                ))
-                .build();
-
-        List<CoordinatorRecord> expectedRecordsBeforeRecomputedAssignments = List.of(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentTombstoneRecord(groupId, oldStaticMemberId),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentTombstoneRecord(groupId, oldStaticMemberId),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMemberTombstoneRecord(groupId, oldStaticMemberId),
-                
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMemberRecord(groupId, expectedCopiedStaticMember),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentRecord(groupId, rejoinStaticMemberId, oldStaticTargetAssignment),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, expectedCopiedStaticMember),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMemberRecord(groupId, expectedUpdatedStaticMember),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMetadataRecord(groupId, bumpedGroupEpoch, groupMetadataHash, 0, getDefaultAssignmentConfigs()
-                )
-        );
-
-        List<CoordinatorRecord> expectedRecomputedTargetAssignmentRecords = List.of(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentRecord(groupId, rejoinStaticMemberId, newStaticTargetAssignment),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentRecord(groupId, dynamicMemberId, newDynamicTargetAssignment)
-        );
-
-        List<CoordinatorRecord> expectedRecordsAfterRecomputedAssignments = List.of(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentMetadataRecord(groupId, bumpedGroupEpoch, context.time.milliseconds()),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, expectedReconciledStaticMember)
-        );
-
-        assertRecordsEquals(
-                expectedRecordsBeforeRecomputedAssignments,
-                rejoinResult.records().subList(0, 8)
-        );
-        assertUnorderedRecordsEquals(
-                List.of(expectedRecomputedTargetAssignmentRecords),
-                rejoinResult.records().subList(8, 10)
-        );
-        assertRecordsEquals(
-                expectedRecordsAfterRecomputedAssignments,
-                rejoinResult.records().subList(10, 12)
-        );
-
-        // WHEN 2: dynamic member send a heartbeat and reconciles to the new target assignment.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> dynamicHeartbeatResult =
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setMemberId(dynamicMemberId)
-                                .setMemberEpoch(groupEpoch)
-                );
-
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(dynamicMemberId)
-                        .setMemberEpoch(bumpedGroupEpoch)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(mkResponseTasks(subtopologyId, 1, 2, 3))
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of()),
-                dynamicHeartbeatResult.response().data()
-        );
-
-        StreamsGroupMember expectedUpdatedDynamicMember = streamsGroupMemberBuilderWithDefaults(dynamicMemberId)
-                .setMemberEpoch(bumpedGroupEpoch)
-                .setPreviousMemberEpoch(groupEpoch)
-                .setAssignedTasks(mkTasksTupleWithEpochs(
-                        TaskRole.ACTIVE,
-                        mkTasksWithEpochs(subtopologyId, Map.of(
-                                1, bumpedGroupEpoch,
-                                2, groupEpoch,
-                                3, groupEpoch
-                        ))
-                ))
-                .build();
-
-        List<CoordinatorRecord> expectedRecordsAfterDynamicHeartbeat = List.of(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, expectedUpdatedDynamicMember)
-        );
-        assertRecordsEquals(expectedRecordsAfterDynamicHeartbeat, dynamicHeartbeatResult.records());
-
-        assertEquals(expectedUpdatedDynamicMember.assignedTasks(), group.getMemberOrThrow(dynamicMemberId).assignedTasks());
-        assertEquals(bumpedGroupEpoch, group.getMemberOrThrow(dynamicMemberId).memberEpoch());
-        assertEquals(newDynamicTargetAssignment, group.targetAssignment(dynamicMemberId, Optional.empty()));
-    }
-
-    @Test
-    public void testStaticRejoinWithSameProcessIdDoesNotBumpEpochAndDynamicHeartbeatRemainsNoOpInMixedGroup() {
-        // Scenario:
-        // There are 2 members. 
-        //   - static member left temporarily with -2 epoch.
-        //   - dynamic member alive in group.
-        // When the same static instance rejoins with the same processId, the coordinator
-        // does not bump the group epoch or recompute the target assignment.
-        // When the dynamic member sends a heartbeat, there is still no new assignment.
-
-        int groupEpoch = 10;
-        String groupId = "fooup";
-
-        String subtopologyId = "subtopology-1";
-        String topicName = "foo";
-        Uuid topicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopologyId).setSourceTopics(List.of(topicName))
-        ));
-
-        String oldStaticMemberId = Uuid.randomUuid().toString();
-        String newStaticMemberId = Uuid.randomUuid().toString();
-        String staticInstanceId = Uuid.randomUuid().toString();
-        String dynamicMemberId = Uuid.randomUuid().toString();
-
-        String staticProcessId = "static-process-id";
-        String dynamicProcessId = "dynamic-process-id";
-
-        // GIVEN Tasks
-        Map.Entry<String, Set<Integer>> staticTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId, 0, 1);
-        TasksTupleWithEpochs staticAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, staticTaskEntries);
-        TasksTuple staticTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, staticTaskEntries);
-
-        Map.Entry<String, Set<Integer>> dynamicTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId, 2, 3);
-        TasksTupleWithEpochs dynamicAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, dynamicTaskEntries);
-        TasksTuple dynamicTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, dynamicTaskEntries);
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(topicId, topicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                topicName, computeTopicHash(topicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(oldStaticMemberId, staticInstanceId)
-                                .setProcessId(staticProcessId)
-                                .setMemberEpoch(StreamsGroupHeartbeatRequest.LEAVE_GROUP_STATIC_MEMBER_EPOCH)
-                                .setPreviousMemberEpoch(groupEpoch)
-                                .setAssignedTasks(staticAssignedTasks)
-                                .build())
-                        .withMember(streamsGroupMemberBuilderWithDefaults(dynamicMemberId)
-                                .setProcessId(dynamicProcessId)
-                                .setMemberEpoch(groupEpoch)
-                                .setPreviousMemberEpoch(groupEpoch - 1)
-                                .setAssignedTasks(dynamicAssignedTasks)
-                                .build())
-                        .withTargetAssignment(oldStaticMemberId, staticTargetAssignment)
-                        .withTargetAssignment(dynamicMemberId, dynamicTargetAssignment)
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs()))
-                .build();
-
-        // WHEN1 : static member try to rejoin with same process id.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> rejoinResult =
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setInstanceId(staticInstanceId)
-                                .setMemberId(newStaticMemberId)
-                                .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-                                .setProcessId(staticProcessId)
-                                .setActiveTasks(List.of())
-                                .setStandbyTasks(List.of())
-                                .setWarmupTasks(List.of())
-                );
-
-        // THEN1
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(newStaticMemberId)
-                        .setMemberEpoch(groupEpoch)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(mkResponseTasks(subtopologyId, 0, 1))
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of()),
-                rejoinResult.response().data()
-        );
-
-        StreamsGroup group = context.groupMetadataManager.streamsGroup(groupId);
-        
-        // no group epoch bump up.
-        assertEquals(groupEpoch, group.groupEpoch());
-        assertFalse(group.hasMember(oldStaticMemberId));
-        assertTrue(group.hasMember(newStaticMemberId));
-        assertTrue(group.hasMember(dynamicMemberId));
-        assertEquals(newStaticMemberId, group.staticMember(staticInstanceId).memberId());
-        assertEquals(staticTargetAssignment, group.targetAssignment(newStaticMemberId, Optional.of(staticInstanceId)));
-        assertEquals(dynamicTargetAssignment, group.targetAssignment(dynamicMemberId, Optional.empty()));
-
-        StreamsGroupMember expectedCopiedStaticMember = streamsGroupMemberBuilderWithDefaults(newStaticMemberId, staticInstanceId)
-                .setProcessId(staticProcessId)
-                .setMemberEpoch(0)
-                .setPreviousMemberEpoch(0)
-                .setAssignedTasks(staticAssignedTasks)
-                .build();
-
-        StreamsGroupMember expectedRejoinedStaticMember = streamsGroupMemberBuilderWithDefaults(newStaticMemberId, staticInstanceId)
-                .setProcessId(staticProcessId)
-                .setMemberEpoch(groupEpoch)
-                .setPreviousMemberEpoch(0)
-                .setAssignedTasks(staticAssignedTasks)
-                .build();
-
-        // no new target assignment.
-        List<CoordinatorRecord> expectedRejoinRecords = List.of(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentTombstoneRecord(groupId, oldStaticMemberId),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentTombstoneRecord(groupId, oldStaticMemberId),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMemberTombstoneRecord(groupId, oldStaticMemberId),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMemberRecord(groupId, expectedCopiedStaticMember),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentRecord(groupId, newStaticMemberId, staticTargetAssignment),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, expectedCopiedStaticMember),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, expectedRejoinedStaticMember)
-        );
-
-        assertRecordsEquals(expectedRejoinRecords, rejoinResult.records());
-
-        // WHEN2 - dynamic member send a heartbeat request
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> dynamicHeartbeatResult =
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setMemberId(dynamicMemberId)
-                                .setMemberEpoch(groupEpoch)
-                );
-
-        // THEN2 - no new target assignment.
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(dynamicMemberId)
-                        .setMemberEpoch(groupEpoch)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(null)
-                        .setStandbyTasks(null)
-                        .setWarmupTasks(null),
-                dynamicHeartbeatResult.response().data()
-        );
-        assertTrue(dynamicHeartbeatResult.records().isEmpty());
-
-        assertEquals(dynamicAssignedTasks, group.getMemberOrThrow(dynamicMemberId).assignedTasks());
-        assertEquals(groupEpoch, group.getMemberOrThrow(dynamicMemberId).memberEpoch());
-        assertEquals(dynamicTargetAssignment, group.targetAssignment(dynamicMemberId, Optional.empty()));
-    }
-
-    @Test
-    public void testOldStaticMemberIdIsFencedAfterReplacementInMixedGroup() {
-        // Scenario:
-        // There are 2 members.
-        //   - static member left temporarily with -2 epoch.
-        //   - dynamic member alive in group.
-        // When the static member rejoins with a new memberId, the old memberId is fenced.
-
-        int groupEpoch = 10;
-        String groupId = "fooup";
-
-        String subtopologyId = "subtopology-1";
-        String topicName = "foo";
-        Uuid topicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopologyId).setSourceTopics(List.of(topicName))
-        ));
-
-        String oldStaticMemberId = Uuid.randomUuid().toString();
-        String newStaticMemberId = Uuid.randomUuid().toString();
-        String staticInstanceId = Uuid.randomUuid().toString();
-        String dynamicMemberId = Uuid.randomUuid().toString();
-
-        String staticProcessId = "static-process-id";
-        String dynamicProcessId = "dynamic-process-id";
-
-        Map.Entry<String, Set<Integer>> staticTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId, 0, 1);
-        TasksTupleWithEpochs staticAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, staticTaskEntries);
-        TasksTuple staticTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, staticTaskEntries);
-
-        Map.Entry<String, Set<Integer>> dynamicTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId, 2, 3);
-        TasksTupleWithEpochs dynamicAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, dynamicTaskEntries);
-        TasksTuple dynamicTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, dynamicTaskEntries);
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(topicId, topicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                topicName, computeTopicHash(topicName, metadataImage)
-        ));
-
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(new MockTaskAssignor("sticky")))
-                .withMetadataImage(metadataImage)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(oldStaticMemberId, staticInstanceId)
-                                .setProcessId(staticProcessId)
-                                .setMemberEpoch(StreamsGroupHeartbeatRequest.LEAVE_GROUP_STATIC_MEMBER_EPOCH)
-                                .setPreviousMemberEpoch(groupEpoch)
-                                .setAssignedTasks(staticAssignedTasks)
-                                .build())
-                        .withMember(streamsGroupMemberBuilderWithDefaults(dynamicMemberId)
-                                .setProcessId(dynamicProcessId)
-                                .setMemberEpoch(groupEpoch)
-                                .setPreviousMemberEpoch(groupEpoch - 1)
-                                .setAssignedTasks(dynamicAssignedTasks)
-                                .build())
-                        .withTargetAssignment(oldStaticMemberId, staticTargetAssignment)
-                        .withTargetAssignment(dynamicMemberId, dynamicTargetAssignment)
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs()))
-                .build();
-
-        // WHEN1 - static member try to rejoin with new member id.
-        context.streamsGroupHeartbeat(
-                new StreamsGroupHeartbeatRequestData()
-                        .setGroupId(groupId)
-                        .setInstanceId(staticInstanceId)
-                        .setMemberId(newStaticMemberId)
-                        .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-                        .setProcessId(staticProcessId)
-                        .setActiveTasks(List.of())
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of())
-        );
-
-        // WHEN2 + THEN2 - stale static member send a heartbeat with stale member id. 
-        assertThrows(FencedInstanceIdException.class, () ->
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setInstanceId(staticInstanceId)
-                                .setMemberId(oldStaticMemberId)
-                                .setMemberEpoch(groupEpoch)
-                )
-        );
-
-        StreamsGroup group = context.groupMetadataManager.streamsGroup(groupId);
-        assertFalse(group.hasMember(oldStaticMemberId));
-        assertTrue(group.hasMember(newStaticMemberId));
-        assertTrue(group.hasMember(dynamicMemberId));
-        assertEquals(newStaticMemberId, group.staticMember(staticInstanceId).memberId());
-    }
-
-    @Test
-    public void testDynamicJoinSucceedsAfterTemporarilyLeftStaticMemberSessionTimeoutInMixedGroup() {
-        // Scenario:
-        // STREAMS_GROUP_MAX_SIZE_CONFIG is 2.
-        // There are 2 members. (1 dynamic member, 1 static member)
-        //   - static member leaves temporarily with -2 epoch.
-        //   - dynamic member alive in group.
-        // After the static member session timeout expires, a new dynamic member can join.
-
-        int groupEpoch = 10;
-        int timeoutGroupEpoch = groupEpoch + 1;
-        int joinGroupEpoch = timeoutGroupEpoch + 1;
-
-        String groupId = "fooup";
-
-        String subtopologyId = "subtopology-1";
-        String topicName = "foo";
-        Uuid topicId = Uuid.randomUuid();
-        Topology topology = new Topology().setSubtopologies(List.of(
-                new Subtopology().setSubtopologyId(subtopologyId).setSourceTopics(List.of(topicName))
-        ));
-
-        String staticMemberId = Uuid.randomUuid().toString();
-        String staticInstanceId = Uuid.randomUuid().toString();
-        String dynamicMemberId = Uuid.randomUuid().toString();
-        String newDynamicMemberId = Uuid.randomUuid().toString();
-
-        String staticProcessId = "static-process-id";
-        String dynamicProcessId = "dynamic-process-id";
-        String newDynamicProcessId = "new-dynamic-process-id";
-
-        // GIVEN assignment
-        Map.Entry<String, Set<Integer>> staticTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId, 0, 1);
-        TasksTupleWithEpochs staticAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, staticTaskEntries);
-        TasksTuple staticTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, staticTaskEntries);
-
-        Map.Entry<String, Set<Integer>> dynamicTaskEntries = TaskAssignmentTestUtil.mkTasks(subtopologyId, 2, 3);
-        TasksTupleWithEpochs dynamicAssignedTasks = mkTasksTupleWithCommonEpoch(TaskRole.ACTIVE, groupEpoch, dynamicTaskEntries);
-        TasksTuple dynamicTargetAssignment = TaskAssignmentTestUtil.mkTasksTuple(TaskRole.ACTIVE, dynamicTaskEntries);
-
-        CoordinatorMetadataImage metadataImage = new MetadataImageBuilder()
-                .addTopic(topicId, topicName, 4)
-                .buildCoordinatorMetadataImage();
-        long groupMetadataHash = computeGroupHash(Map.of(
-                topicName, computeTopicHash(topicName, metadataImage)
-        ));
-
-        MockTaskAssignor assignor = new MockTaskAssignor("sticky");
-        GroupMetadataManagerTestContext context = new GroupMetadataManagerTestContext.Builder()
-                .withStreamsGroupTaskAssignors(List.of(assignor))
-                .withMetadataImage(metadataImage)
-                .withConfig(GroupCoordinatorConfig.STREAMS_GROUP_MAX_SIZE_CONFIG, 2)
-                .withStreamsGroup(new StreamsGroupBuilder(groupId, groupEpoch)
-                        .withMember(streamsGroupMemberBuilderWithDefaults(staticMemberId, staticInstanceId)
-                                .setProcessId(staticProcessId)
-                                .setMemberEpoch(groupEpoch)
-                                .setPreviousMemberEpoch(groupEpoch - 1)
-                                .setAssignedTasks(staticAssignedTasks)
-                                .build())
-                        .withMember(streamsGroupMemberBuilderWithDefaults(dynamicMemberId)
-                                .setProcessId(dynamicProcessId)
-                                .setMemberEpoch(groupEpoch)
-                                .setPreviousMemberEpoch(groupEpoch - 1)
-                                .setAssignedTasks(dynamicAssignedTasks)
-                                .build())
-                        .withTargetAssignment(staticMemberId, staticTargetAssignment)
-                        .withTargetAssignment(dynamicMemberId, dynamicTargetAssignment)
-                        .withTargetAssignmentEpoch(groupEpoch)
-                        .withTopology(StreamsTopology.fromHeartbeatRequest(topology))
-                        .withValidatedTopologyEpoch(0)
-                        .withMetadataHash(groupMetadataHash)
-                        .withLastAssignmentConfigs(getDefaultAssignmentConfigs()))
-                .build();
-        context.onLoaded();
-
-        // WHEN1 - static member leaves with epoch -2.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> leaveResult =
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setInstanceId(staticInstanceId)
-                                .setMemberId(staticMemberId)
-                                .setMemberEpoch(StreamsGroupHeartbeatRequest.LEAVE_GROUP_STATIC_MEMBER_EPOCH)
-                );
-
-        // THEN1
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(staticMemberId)
-                        .setMemberEpoch(StreamsGroupHeartbeatRequest.LEAVE_GROUP_STATIC_MEMBER_EPOCH)
-                        .setActiveTasks(null)
-                        .setWarmupTasks(null)
-                        .setStandbyTasks(null)
-                        .setStatus(List.of()),
-                leaveResult.response().data()
-        );
-
-        // To prevent session timeout from dynamic member.
-        // Sleep 1, and dynamic member send a heartbeat.
-        GroupMetadataManagerTestContext.assertNoOrEmptyResult(context.sleep(1));
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> dynamicHeartbeatResult =
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setMemberId(dynamicMemberId)
-                                .setMemberEpoch(groupEpoch)
-                );
-
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(dynamicMemberId)
-                        .setMemberEpoch(groupEpoch)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(null)
-                        .setStandbyTasks(null)
-                        .setWarmupTasks(null),
-                dynamicHeartbeatResult.response().data()
-        );
-        assertTrue(dynamicHeartbeatResult.records().isEmpty());
-        
-        // WHEN2: static member session timeout.
-        context.assertSessionTimeout(groupId, staticMemberId, 45000 - 1);
-        List<ExpiredTimeout<CoordinatorRecord>> timeouts = context.sleep(45000 - 1);
-
-        // THEN2
-        List<CoordinatorRecord> expectedTimeoutRecords = List.of(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentTombstoneRecord(groupId, staticMemberId),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentTombstoneRecord(groupId, staticMemberId),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMemberTombstoneRecord(groupId, staticMemberId),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMetadataRecord(
-                        groupId, timeoutGroupEpoch, groupMetadataHash, 0,getDefaultAssignmentConfigs()
-                )
-        );
-
-        assertEquals(
-                List.of(new ExpiredTimeout<>(
-                        groupSessionTimeoutKey(groupId, staticMemberId),
-                        new CoordinatorResult<>(expectedTimeoutRecords)
-                )),
-                timeouts
-        );
-
-        StreamsGroup group = context.groupMetadataManager.streamsGroup(groupId);
-        assertFalse(group.hasMember(staticMemberId));
-        assertTrue(group.hasMember(dynamicMemberId));
-        assertEquals(timeoutGroupEpoch, group.groupEpoch());
-
-        assignor.prepareGroupAssignment(Map.of(
-                dynamicMemberId, dynamicTargetAssignment,
-                newDynamicMemberId, staticTargetAssignment
-        ));
-
-        // WHEN3 - new dynamic member try to join.
-        CoordinatorResult<StreamsGroupHeartbeatResult, CoordinatorRecord> joinResult =
-                context.streamsGroupHeartbeat(
-                        new StreamsGroupHeartbeatRequestData()
-                                .setGroupId(groupId)
-                                .setMemberId(newDynamicMemberId)
-                                .setMemberEpoch(StreamsGroupHeartbeatRequest.JOIN_GROUP_MEMBER_EPOCH)
-                                .setProcessId(newDynamicProcessId)
-                                .setRebalanceTimeoutMs(1500)
-                                .setTopology(topology)
-                                .setActiveTasks(List.of())
-                                .setStandbyTasks(List.of())
-                                .setWarmupTasks(List.of())
-                );
-
-        // THEN3 : accept join.
-        assertResponseEquals(
-                new StreamsGroupHeartbeatResponseData()
-                        .setMemberId(newDynamicMemberId)
-                        .setMemberEpoch(joinGroupEpoch)
-                        .setHeartbeatIntervalMs(5000)
-                        .setTaskOffsetIntervalMs(60000)
-                        .setActiveTasks(mkResponseTasks(subtopologyId, 0, 1))
-                        .setStandbyTasks(List.of())
-                        .setWarmupTasks(List.of()),
-                joinResult.response().data()
-        );
-
-        StreamsGroupMember expectedJoiningDynamicMember = streamsGroupMemberBuilderWithDefaults(newDynamicMemberId)
-                .setProcessId(newDynamicProcessId)
-                .setMemberEpoch(0)
-                .setPreviousMemberEpoch(0)
-                .build();
-
-        StreamsGroupMember expectedReconciledDynamicMember = streamsGroupMemberBuilderWithDefaults(newDynamicMemberId)
-                .setProcessId(newDynamicProcessId)
-                .setMemberEpoch(joinGroupEpoch)
-                .setPreviousMemberEpoch(0)
-                .setAssignedTasks(mkTasksTupleWithEpochs(
-                        TaskRole.ACTIVE,
-                        mkTasksWithEpochs(subtopologyId, Map.of(
-                                0, joinGroupEpoch,
-                                1, joinGroupEpoch
-                        ))
-                ))
-                .build();
-
-        List<CoordinatorRecord> expectedJoinRecords = List.of(
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMemberRecord(groupId, expectedJoiningDynamicMember),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupMetadataRecord(
-                        groupId,
-                        joinGroupEpoch,
-                        groupMetadataHash,
-                        0,
-                        getDefaultAssignmentConfigs()
-                ),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentRecord(groupId, newDynamicMemberId, staticTargetAssignment),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupTargetAssignmentMetadataRecord(groupId, joinGroupEpoch, context.time.milliseconds()),
-                StreamsCoordinatorRecordHelpers.newStreamsGroupCurrentAssignmentRecord(groupId, expectedReconciledDynamicMember)
-        );
-
-        assertRecordsEquals(expectedJoinRecords, joinResult.records());
-
-        assertTrue(group.hasMember(dynamicMemberId));
-        assertTrue(group.hasMember(newDynamicMemberId));
-        assertEquals(joinGroupEpoch, group.groupEpoch());
-        assertEquals(staticTargetAssignment, group.targetAssignment(newDynamicMemberId, Optional.empty()));
-        assertEquals(dynamicTargetAssignment, group.targetAssignment(dynamicMemberId, Optional.empty()));
-    }
-
-    @Test
     public void testStreamsLeavingMemberRemovesMemberAndBumpsGroupEpoch() {
         String groupId = "fooup";
         String memberId1 = Uuid.randomUuid().toString();
@@ -30842,14 +28414,4 @@ public class GroupMetadataManagerTest {
         ));
     }
 
-    private static List<StreamsGroupHeartbeatResponseData.TaskIds> mkResponseTasks(
-            String subtopologyId,
-            Integer... partitions
-    ) {
-        return List.of(
-                new StreamsGroupHeartbeatResponseData.TaskIds()
-                        .setSubtopologyId(subtopologyId)
-                        .setPartitions(Arrays.asList(partitions))
-        );
-    }
 }
